webapp/requirements.txt
Flask 
docling 
qdrant-client 
sentence-transformers 
langchain 
python-dotenv
groq
docling
langchain-community